Removed useless nullability and toDouble

+++ ionav/src/main/java/de/ironjan/arionav_fw/ionav/IonavMapView.kt
@@ -6,8 +6,8 @@ import android.widget.Toast
 import androidx.lifecycle.LifecycleOwner
 import androidx.lifecycle.Observer
 import de.ironjan.arionav_fw.ionav.custom_view_mvvm.MvvmCustomView
-import de.ironjan.arionav_fw.ionav.views.mapview.*
 import de.ironjan.arionav_fw.ionav.routing.RoutingService
+import de.ironjan.arionav_fw.ionav.views.mapview.*
 import de.ironjan.graphhopper.extensions_core.Coordinate
 import org.oscim.android.MapView
 import org.oscim.core.GeoPoint
@@ -155,7 +155,7 @@ class IonavMapView : MapView, MvvmCustomView<IonavViewModel> {
         if (viewModel.routingStatus.value == RoutingService.Status.READY) {
             val selectedLevel = viewModel.getSelectedLevel()
 
-            viewModel.setDestination(Coordinate(p.latitude, p.longitude, selectedLevel.toDouble()))
+            viewModel.setDestination(Coordinate(p.latitude, p.longitude, selectedLevel))
             return true
         }
